malte/requirements.txt
Django==2.2.17
django-cors-headers==3.5.0
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
httplib2==0.9.1
matplotlib==3.0.3
numpy==1.18.5
oauthlib==1.0.3
Pillow==3.1.2
psycopg2==2.8.6
ptyprocess==0.5
pyasn1==0.1.9
pycrypto==2.6.1
pycups==1.9.73
PyJWT==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
spark-parser==1.8.9
sqlparse==0.4.1
urllib3==1.13.1
whitenoise==5.2.0
opencv-python==4.4.0.42